Rename ssh.h to not conflict with openssl/ssl.h.
